requirements.txt
PySide6>=6.7.0
pyinstaller>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
imagehash>=4.3.0
tqdm>=4.65.0
